Replace invalid storeURL with cacheURL to properly implement the Non permission deletion error

+++ EssentialFeed/Feed Cache/CodableFeedStore.swift
@@ -72,6 +72,10 @@ public class CodableFeedStore : FeedStore {
     }
     
     public func deleteCachedFeed(completion :@escaping DeletionCompletion){
+        guard FileManager.default.fileExists(atPath: storeURL.path) else {
+           return completion(nil)
+        }
+        
         do {
             try FileManager.default.removeItem(at: storeURL)
             completion(nil)
